FacebookData (#10)
* Get facebook reactions

Given the facebook page of a news source and a link to an article, finds
a post on the page that contains the given link(if any), and returns the
total count for each type of reaction.

* Updated fb post search to work for most urls

The usually shortened urls in the facebook posts are first converted to
the long url that they finally direct to, prior to being compared to the
input url. The comparison is only based on the hierarchical paths in the
two urls.

* Search multiple result pages

Program now receives an extra argument, specifying how far in the list
of posts to go with the search.

* Request posts from a given timeframe

Specify timeframe for posts in query. Changed from using facebook-sdk to
using the requests package directly for accessing the graph api. Also
stop after finding a matching url.

* Fixed pagination issue

Plus preventing crash if trying to unshorten badly formatted url

* style

* Removed limit on the number of redirects